Clean and fix

+++ src/ConfigFiles.h
@@ -175,7 +175,7 @@ public:
 		}
 		else
 		{
-			std::cerr << "ConfigVector::get(" << id << ") : warning : id is not valid, returning default" << std::endl;
+			std::cerr << "ConfigVector::get(" << static_cast<unsigned int>(id) << ") : warning : id is not valid, returning default" << std::endl;
 			return &defaultEntry;
 		}
 	}
